+++ rollup-common.js
@@ -1,8 +1,8 @@
 import json from '@rollup/plugin-json';
+import { importMetaAssets } from '@web/rollup-plugin-import-meta-assets';
 import glob from 'glob';
 import babel from 'rollup-plugin-babel';
 import clear from 'rollup-plugin-clear';
-import copy from 'rollup-plugin-copy';
 import { terser } from 'rollup-plugin-terser';
 import visualizer from 'rollup-plugin-visualizer';
 import SVGO from 'svgo';
@@ -94,16 +94,14 @@ export function plugins (sourceDir, outputDir) {
       targets: [outputDir],
     }),
     json(),
-    copy({
-      targets: [{
-        src: `${sourceDir}/assets/*.svg`,
-        dest: outputDir + `/assets`,
-        transform: (svgBuffer) => {
-          return svgo
-            .optimize(svgBuffer.toString())
-            .then(({ data }) => data);
-        },
-      }],
+    importMetaAssets({
+      // Let's assume we don't have import.meta.url assets in our deps to speed up things
+      exclude: 'node_modules/**',
+      transform: (svgBuffer, id) => {
+        return svgo
+          .optimize(svgBuffer.toString())
+          .then(({ data }) => data);
+      },
     }),
     terser({
       output: { comments: false },