+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### For devs
 
+* Use `@web/rollup-plugin-import-meta-assets` for assets instead of `rollup-plugin-copy`
 * Replace `assetUrl()` helper with raw `new URL('../asset.svg', import.meta.url).href`
 
 ## 5.3.1 (2020-11-16)